requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
redis==5.0.8
pydantic==2.8.2
python-dotenv==1.0.1
